move privatedns tests
