+++ azurerm/internal/services/privatedns/private_dns_cname_record_resource_test.go
@@ -1,4 +1,4 @@
-package tests
+package privatedns_test
 
 import (
 	"fmt"
