+++ azurerm/internal/services/privatedns/private_dns_srv_record_resource_test.go
@@ -1,4 +1,4 @@
-package tests
+package privatedns_test
 
 import (
 	"fmt"
