+++ azurerm/internal/services/privatedns/private_dns_zone_data_source_test.go
@@ -1,4 +1,4 @@
-package tests
+package privatedns_test
 
 import (
 	"fmt"
